P4-2434 removing now redundant code and some refactorings to boot.
