+++ helm_deploy/calculate-journey-variable-payments/templates/secrets.yaml
@@ -9,4 +9,3 @@ data:
   BASM_AWS_SECRET_ACCESS_KEY: {{ .Values.secrets.BASM_AWS_SECRET_ACCESS_KEY | b64enc | quote }}
   HMPPS_AUTH_CLIENT_ID: {{ .Values.secrets.HMPPS_AUTH_CLIENT_ID | b64enc | quote }}
   HMPPS_AUTH_CLIENT_SECRET: {{ .Values.secrets.HMPPS_AUTH_CLIENT_SECRET | b64enc | quote }}
-  SPREADSHEET_PASSWORD: {{ .Values.secrets.SPREADSHEET_PASSWORD | b64enc | quote }}
