+++ helm_deploy/secrets-example.yaml
@@ -5,4 +5,3 @@ secrets:
   BASM_AWS_SECRET_ACCESS_KEY: somesecretvalue
   HMPPS_AUTH_CLIENT_ID: somesecretvalue
   HMPPS_AUTH_CLIENT_SECRET: somesecretvalue
-  SPREADSHEET_PASSWORD: somesecretvalue
