+++ src/main/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/import/InboundSpreadsheet.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.pecs.jpc
+package uk.gov.justice.digital.hmpps.pecs.jpc.import
 
 import org.apache.poi.ss.usermodel.Row
 import org.apache.poi.ss.usermodel.Workbook
