--- src/main/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/service/ImportStatus.kt
@@ -1,6 +1,0 @@
-package uk.gov.justice.digital.hmpps.pecs.jpc.service
-
-enum class ImportStatus() {
-    DONE,
-    IN_PROGRESS
-}
