--- src/main/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/spreadsheet/SpreadsheetProtection.kt
@@ -1,7 +1,0 @@
-package uk.gov.justice.digital.hmpps.pecs.jpc.spreadsheet
-
-import java.io.File
-
-fun interface SpreadsheetProtection {
-    fun protectAndGet(file: File): File
-}
