--- src/main/resources/exampleApplication.yml
@@ -1,58 +1,0 @@
-info.app:
-  name: Calculate Journey Variable Payments
-  version: 1.0
-
-spring:
-  application:
-    name: calculate-journey-variable-payments
-  codec:
-    max-in-memory-size: 50MB
-  jpa:
-    hibernate:
-      ddl-auto: create
-    database-platform: org.hibernate.dialect.PostgreSQLDialect
-    datasource:
-      url: "jdbc:postgresql://localhost:5432/jpc"
-      username: bob
-      password:
-
-
-server:
-  port: 8080
-  servlet:
-    context-path: /
-  forward-headers-strategy: native
-  tomcat:
-    remoteip:
-      remote-ip-header: x-forwarded-for
-      protocol-header: x-forwarded-proto
-  shutdown: graceful
-  error:
-    include-message: always
-
-management:
-  endpoints:
-    web:
-      base-path: /
-      exposure:
-        include: 'info, health'
-  endpoint:
-    health:
-      cache:
-        time-to-live: 2000ms
-      show-components: always
-      show-details: always
-      probes:
-        enabled: true
-    info:
-      cache:
-        time-to-live: 2000ms
-logging:
-  level:
-    ROOT: INFO
-import-files:
-  locations: "/locations.xlsx"
-  serco-prices: "/serco_prices.xlsx"
-  geo-prices: "/geo_prices.xlsx"
-
-
